Obliged to move method to avoid circular imports
